Fixed return type of DumpRenderTreeSupportQt::javaScriptObjectsCount().

+++ Source/WebKit/qt/WebCoreSupport/DumpRenderTreeSupportQt.cpp
@@ -272,7 +272,7 @@ void DumpRenderTreeSupportQt::clearFrameName(QWebFrameAdapter *adapter)
     coreFrame->tree().clearName();
 }
 
-int DumpRenderTreeSupportQt::javaScriptObjectsCount()
+size_t DumpRenderTreeSupportQt::javaScriptObjectsCount()
 {
     return JSDOMWindowBase::commonVM().heap.globalObjectCount();
 }
